Fix type mismatch in invoice status calculations
- Cast Money amounts to integers in InvoiceCalculator to prevent float/int comparison issues
- Ensure Money constructor always stores amounts as integers
- Fixes isPaid() returning false due to strict comparison (===) between float and int

Agent-Logs-Url: https://github.com/Bottelet/DaybydayCRM/sessions/0f9f225c-1173-42cb-8b98-1f65c799b97e

+++ app/Services/Invoice/InvoiceCalculator.php
@@ -33,7 +33,7 @@ class InvoiceCalculator
     {
         $subTotal = $this->getSubTotal()->getAmount();
 
-        return new Money($subTotal * $this->tax->vatRate());
+        return new Money((int) ($subTotal * $this->tax->vatRate()));
     }
 
     public function getTotalPrice(): Money
@@ -41,7 +41,7 @@ class InvoiceCalculator
         $subTotal = $this->getSubTotal()->getAmount();
         $vatTotal = $this->getVatTotal()->getAmount();
 
-        return new Money($subTotal + $vatTotal);
+        return new Money((int) ($subTotal + $vatTotal));
     }
 
     public function getSubTotal(): Money
